modiff

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: madamou <madamou@contact.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/19 15:06:46 by marvin            #+#    #+#             */
-/*   Updated: 2024/03/23 17:28:14 by madamou          ###   ########.fr       */
+/*   Updated: 2024/03/30 19:09:53 by madamou          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	i;
